// src_old/app/store/content/types/positions/PositionsActionTypes.enum.ts
export enum EPositionsActionTypes {
  FIND_ALL = '[Positions]-find-all',
  FIND_BY_ID = '[Positions]-find-by-id',
  CREATE = '[Positions]-create',
  DELETE = '[Positions]-delete',
  UPDATE = '[Positions]-update',
  SUCCESS_GET_ONE = '[Positions]-get-one-success',
  SUCCESS_GET_MANY = '[Positions]-get-many-success',
  SUCCESS_DELETE = '[Positions]-delete-success',
  ERROR = '[Positions]-failure',
}